Complete mongodb query to api
